add: 재사용을 위해 AddIngredientButton을 BorderImageButton으로 변경
centerImage 설정을 통해 아이콘 변경 가능하도록 구현
