+++ Reciptopia/Presenters+Dependencies/CheckIngredient/CheckIngredientViewController.storyboard
@@ -106,7 +106,7 @@
                                     <outlet property="delegate" destination="Y6W-OH-hqX" id="pmh-qJ-knn"/>
                                 </connections>
                             </collectionView>
-                            <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="hkT-K1-kNu" customClass="AddIngredientButton" customModule="Reciptopia" customModuleProvider="target">
+                            <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="hkT-K1-kNu" customClass="BorderImageButton" customModule="Reciptopia" customModuleProvider="target">
                                 <rect key="frame" x="50" y="291" width="314" height="45"/>
                                 <accessibility key="accessibilityConfiguration" identifier="addIngredientButton"/>
                                 <constraints>
@@ -121,6 +121,7 @@
                                     <userDefinedRuntimeAttribute type="number" keyPath="borderWidth">
                                         <real key="value" value="1"/>
                                     </userDefinedRuntimeAttribute>
+                                    <userDefinedRuntimeAttribute type="image" keyPath="centerImage" value="plus" catalog="system"/>
                                 </userDefinedRuntimeAttributes>
                                 <connections>
                                     <action selector="onAddIngredient:" destination="Y6W-OH-hqX" eventType="touchUpInside" id="FV4-Mk-ElX"/>
@@ -187,6 +188,7 @@
         </designable>
     </designables>
     <resources>
+        <image name="plus" catalog="system" width="128" height="113"/>
         <image name="xmark" catalog="system" width="128" height="113"/>
         <namedColor name="AccentColor">
             <color red="0.37299999594688416" green="0.56099998950958252" blue="0.92900002002716064" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
